Moved service get-jobs into its own folder
