+++ src/app/Components/job-carousel/job-carousel.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable, of, from, map } from 'rxjs';
 import { Job } from '../../Interfaces/Company';
-import { GetJobsService } from '../../Services/get-jobs.service';
+import { GetJobsService } from '../../Services/get-jobs/get-jobs.service';
 import { TechStack } from '../../Interfaces/Company';
 
 @Component({
