+++ src/app/Services/get-jobs/get-jobs.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Job } from '../Interfaces/Company';
+import { Job } from '../../Interfaces/Company';
 
 @Injectable({
   providedIn: 'root'
